Move ConnectionSetup to scalecube-services-api module
