+++ services-transport-parent/services-transport-rsocket/src/main/java/io/scalecube/services/transport/rsocket/RSocketClientTransport.java
@@ -10,6 +10,8 @@ import io.rsocket.frame.decoder.PayloadDecoder;
 import io.rsocket.util.ByteBufPayload;
 import io.scalecube.net.Address;
 import io.scalecube.services.ServiceReference;
+import io.scalecube.services.auth.ConnectionSetup;
+import io.scalecube.services.auth.ConnectionSetupCodec;
 import io.scalecube.services.auth.CredentialsSupplier;
 import io.scalecube.services.exceptions.MessageCodecException;
 import io.scalecube.services.exceptions.ServiceException;
